03-rag-system/requirements.txt
# Core dependencies
anthropic>=0.40.0
openai>=1.54.0
python-dotenv>=1.0.0

# Vector stores
chromadb>=0.4.0
# pinecone-client>=3.0.0  # Uncomment for Pinecone

# Embeddings
sentence-transformers>=2.2.2
transformers>=4.30.0

# Retrieval
rank-bm25>=0.2.2

# Document processing
nltk>=3.8.1
spacy>=3.7.0
pypdf>=3.15.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Evaluation
scikit-learn>=1.3.0
